Specifies rubocop target version
Sonatype Lift is using Rubocop to report quality violations.
The default configuration Sonatype Lift is using is not
publically available. So in order to fix the comment that it
left on the pull request after the version requirement was
added to the `.gemspec` file, I added Rubocop as a development
dependency, and then generated a TODO style configuration that
excludes all of the default warnings except for the one that
Sonatype Lift reported. That one was fixed.

Signed-off-by: M. Scott Ford <scott@corgibytes.com>