Bump version to 1.2.0 and upgrade dependencies
`cyclonedx-ruby` has a flag, `-f` or `--format`, that allows for json
output in addition to xml. This functionality has been added to the code
but has not been released. By bumping the version number, a new gem can
be cut that has the enhanced output functionality.

Additionally, the gem's dependencies were updated to newer versions.
Some of these upgrades, such as the ones for `json` and `nokogiri`,
contain bug fixes, but all are compatible with Ruby 2.7.0 and greater,
which is the version of Ruby pegged in the `cyclonedx-ruby.gemspec`
file.

Signed-off-by: Ryan Cammer <ryancammer@gmail.com>